pyproject.toml
[project]
name = "freshservice-mcp"
version = "1.0.0"
description = "An MCP server for Freshservice"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "build>=1.2.2.post1",
    "httpx==0.28.1",
    "mcp[cli]>=1.8.0",
    "pydantic>=2.10.6",
    "prometheus-client>=0.21",
]

[project.optional-dependencies]
redis = ["redis[hiredis]>=5.0.0"]
observability = [
    "opentelemetry-api>=1.27",
    "opentelemetry-sdk>=1.27",
    "opentelemetry-exporter-otlp-proto-grpc>=1.27",
    "opentelemetry-instrumentation-httpx>=0.48b0",
]

[[project.authors]]
name = "Gopi Krishnan"
email = "gopi@effy.co.in"

[[project.authors]]
name = "Maanaesh Swamy"
email = "maanaesh.s@effy.co.in"

[[project.authors]]
name = "Vijay Ragunath"
email = "vijay.r@effy.co.in"

[project.scripts]
freshservice-mcp = "freshservice_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.",
    "raise NotImplementedError",
]
